backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
python-multipart
loguru
